remove CustomMenu from project
finished refactoring out the contents of CustomMenu
