+++ xlive/Blam/Engine/interface/screens/screen_press_start_introduction.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "interface/user_interface.h"
+
+void screen_press_start_introduction_apply_patches();
