+++ xlive/Blam/Engine/interface/screens/screens_patches.cpp
@@ -6,6 +6,7 @@
 #include "screen_4way_signin.h"
 #include "screen_main_menu.h"
 #include "screen_multiplayer_pregame_lobby.h"
+#include "screen_press_start_introduction.h"
 #include "screen_squad_settings.h"
 #include "screen_settings.h"
 
@@ -20,6 +21,8 @@ void screens_apply_patches_on_map_load()
 
 void screens_apply_patches()
 {
+	screen_press_start_introduction_apply_patches();
+
 	c_screen_esrb_warning::apply_patches();
 	c_screen_4way_signin::apply_instance_patches();
 	c_main_menu_list::apply_instance_patches();
