--- xlive/H2MOD/Modules/CustomMenu/CustomMenu.h
@@ -1,10 +1,0 @@
-#pragma once
-
-void InitCustomMenu();
-void DeinitCustomMenu();
-
-void* ui_load_cartographer_invalid_login_token();
-
-extern bool g_force_cartographer_update;
-
-void XUiShowSignInH2();
